Fixes issue #31
jQuery handles events from `.trigger` a bit interesting in that it checks if there's a matching method and runs it instead. Since `.remove` is a jQuery and a dom method, that took priority.

I've namespaced all events we trigger which will break any code that relies on it. It's the only way to make sure anything else funky doesn't happen.